+++ tfhe/src/integer/server_key/mod.rs
@@ -246,7 +246,7 @@ impl ServerKey {
             MaxDegree::from_msg_carry_modulus(self.message_modulus(), self.carry_modulus());
         let max_sum_to_full_carry = max_degree.get() / degree.get();
 
-        max_sum_to_full_carry.min(self.key.max_noise_level.get())
+        max_sum_to_full_carry.min(self.key.max_noise_level.get() as usize)
     }
 }
